python/requirements.txt
pandas>=1.5.0
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
